tidy up the llm provider
decouple a big ball of mud into seperate file per llm provider
