+++ opsmate/dino/provider/__init__.py
@@ -1,0 +1,13 @@
+from .base import Provider, discover_providers, register_provider
+from .openai import OpenAIProvider
+from .anthropic import AnthropicProvider
+from .xai import XAIProvider
+
+__all__ = [
+    "Provider",
+    "OpenAIProvider",
+    "AnthropicProvider",
+    "XAIProvider",
+    "discover_providers",
+    "register_provider",
+]
